+++ src/components/common/Tierlist/TierlistTypes.ts
@@ -1,5 +1,5 @@
-import { LocalStorageReturn } from "@/hooks/useLocalStorage";
 import { ReactNode } from "react";
+import { LocalStorageReturn } from "@/hooks/useLocalStorage";
 
 export interface Entry<TItem> {
   id: string;
