+++ src/components/global/Header/components/SettingsContainer.tsx
@@ -4,7 +4,7 @@ import SettingsCog from "@/components/common/icons/SettingsCog";
 import { CharacterImage } from "@/components/common/Images";
 import { useRegionStore } from "@/stores/RegionStore";
 
-import SettingsOption from "./SettingsModal/SettingsOption";
+import SettingsOption from "../../../../stores/SettingsStore/components/SettingsModal/SettingsOption";
 import { addTabNavigation } from "@/common/functions/accessibility";
 
 const debugLog = DebugLog(DebugLog.DEBUGS.settingsContainer);
