+++ src/components/global/Header/components/index.ts
@@ -2,4 +2,4 @@ export { default as Cache } from './Cache';
 export { default as Search } from './Search';
 export { default as Navigation } from './Navigation';
 export { default as SettingsContainer } from './SettingsContainer';
-export { default as SettingsModal } from './SettingsModal';
+export { default as SettingsModal } from '../../../../stores/SettingsStore/components/SettingsModal';
