+++ src/pages/Building/PriorityList/components/index.ts
@@ -1,2 +1,3 @@
 export * from './CreatePriorityListButton';
-export * from './ModifyPriorityListModal'
+export * from './ModifyPriorityListModal';
+export * from './PriorityListTab';
