+++ src/stores/SettingsStore/components/SettingsModal/SettingsModal.scss
